refactor(admin): split admin.schemas.ts into 6 domain schema files
Break down the 1014-line monolithic admin.schemas.ts into organized
domain files under schemas/admin/ for better readability and maintenance.
Original file becomes a barrel re-export for backward compatibility.
