+++ apps/web/src/lib/api/schemas/admin/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Admin Schemas Index
+ *
+ * Centralized export for all admin domain schema files.
+ */
+
+export * from './admin-users.schemas';
+export * from './admin-content.schemas';
+export * from './admin-ai.schemas';
+export * from './admin-analytics.schemas';
+export * from './admin-system.schemas';
+export * from './admin-monitor.schemas';
